Show movie and can view movie detail

+++ database/migrations/2024_08_19_071259_create_movies_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->text('description');
             $table->string('release_year');
-            $table->integer('duration'); // Duration in minutes
+            $table->double('duration'); // Duration in minutes
             $table->string('quality');
             $table->string('language');
             $table->string('cover_image')->nullable();
